requirements.txt
django==1.11.7
django-debug-toolbar==1.8
gunicorn==19.4.5
psycopg2==2.7.3.1
whitenoise==3.3.1
django-sql-explorer==1.1.1
djangorestframework==3.7.7
mysql-connector-python==8.0.6
